Add ArenaOrderedSga_allocate codegen
Unformat arguments _after_ marshalling the return value into a C type.
This ensures if the lifetime of the return value depends on the lifetime
of an argument, that the pointer to the return value is written first.

Add Ref ArgType which is the same as VoidPtr except when used as an
argument to a function it is used as &arg0 instead of arg0.